fix for some dwc-a specifications
The parser crashed before if the coreid was mapped to a term that was also listed as one of the fields (and was part of the MIDS mapped fields). Now the "id" field is only added if it's not listed in meta.xml or not selected by the SSSOM mapping.